Do not show 'Total' column in resources
This column does not have any sence, because it shows incorrect wall time (as a sum for all components),
incorrect memory usage (it is unclear, if it shows usage of a single runs, several runs on the same machine, several runs on several machines or even of scripts),
and potentially duplicated CPU time.
In any case, it more easy to create some component with overall resources, where they can be computed correctly.